requirements.txt
rich>=14.0.0
textual>=0.6.0
python-docx>=1.1.2
docx2pdf>=0.1.8